skills: break the type-only cycle skills/types.ts ↔ menu-config.ts (#1091)
`types.ts` imported `MenuConfig` (type-only) from `menu-config.ts`, which in
turn imports `SkillMenu`/`SKILL_MENUS` from `types.ts` — a type-only cycle
(erased at compile, so no runtime cycle, but it showed up in madge).

Move the three config shapes (`MenuConfig`, `SkillSetting`, `MenuItemLike`) into
the leaf `types.ts` — they only depend on `SkillMenu`, which already lives
there. `menu-config.ts` keeps the pure logic and imports the types back,
re-exporting them so existing `from './menu-config'` importers are unchanged.

madge now reports no circular dependency across the project (was the only one).
